scene-control plugin
