+++ packages/engine/src/index.ts
@@ -2,6 +2,7 @@ export * from './HedronEngine/HedronEngine'
 export type * from './HedronEngine/types'
 export * from '@plugins/Plugin'
 export * from '@store/types'
+export * from '@store/storeListener'
 export { getNextEnumValue } from '@store/selectors/getNextEnumValue'
 export * from './globalVars'
 export { type EngineStore } from './store/engineStore'
