+++ packages/scene-control/src/components/SceneControlGlobalPanel/SceneControlGlobalPanel.module.css
