+++ packages/scene-control/src/components/SceneControlGlobalWidget/SceneControlGlobalWidget.module.css
@@ -1,0 +1,26 @@
+.sceneControlButton {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border: 1px solid var(--color-border, #3a3a3a);
+  background: var(--color-bg-secondary, #1f1f1f);
+  color: var(--color-text, #f2f2f2);
+  border-radius: 6px;
+  padding: 6px 10px;
+  cursor: pointer;
+  font-size: 12px;
+  line-height: 1;
+}
+
+.sceneControlButton:hover {
+  border-color: var(--color-accent, #7db5ff);
+}
+
+.active {
+  border-color: var(--color-accent, #7db5ff);
+  background: var(--color-bg-tertiary, #2b2b2b);
+}
+
+.label {
+  font-weight: 600;
+}
