+++ packages/scene-control/src/index.ts
@@ -1,0 +1,3 @@
+export { SceneControlPlugin } from './SceneControlPlugin'
+export { SceneControlGlobalWidget } from './components/SceneControlGlobalWidget/SceneControlGlobalWidget'
+export { SceneControlGlobalPanel } from './components/SceneControlGlobalPanel/SceneControlGlobalPanel'
