+++ packages/scene-control/tsconfig.json
@@ -1,0 +1,8 @@
+{
+  "extends": ["../../tsconfig.json"],
+  "include": ["src/**/*"],
+  "compilerOptions": {
+    "jsx": "react-jsx",
+    "baseUrl": "."
+  }
+}
