+++ packages/scene-control/vite.config.ts
@@ -1,0 +1,31 @@
+import { defineConfig } from 'vite'
+import tsconfigPaths from 'vite-tsconfig-paths'
+import react from '@vitejs/plugin-react'
+import dts from 'vite-plugin-dts'
+
+export default defineConfig({
+  plugins: [react(), tsconfigPaths(), dts({ rollupTypes: true })],
+  build: {
+    lib: {
+      entry: ['src/index.ts'],
+      name: 'scene-control',
+      fileName: (format) => `scene-control.${format}.js`,
+    },
+    rollupOptions: {
+      external: [
+        'react',
+        'react-dom',
+        'react/jsx-runtime',
+        '@hedron-gl/engine',
+        '@hedron-gl/ui-core',
+      ],
+      output: {
+        globals: {
+          react: 'React',
+          'react-dom': 'ReactDOM',
+          'react/jsx-runtime': 'React',
+        },
+      },
+    },
+  },
+})
